+++ 2dPlatformEngine/src/fr/spriggans/game/level/levelObjects/background/Background.java
@@ -13,6 +13,6 @@ public class Background extends AbstractLevelElement {
 
 	@Override
 	public void render(Screen screen) {
-		Bitmap.BG_TEST.render(screen, x, y, 0);
+		Bitmap.BG_TEST.render(screen, (int) x, (int) y, 0);
 	}
 }
